+++ src/OmnichannelChatSDK.ts
@@ -332,7 +332,7 @@ class OmnichannelChatSDK {
             }
         }
 
-        if (optionalParams.liveChatContext && !this.reconnectId) {
+        if (optionalParams.liveChatContext && Object.keys(optionalParams.liveChatContext).length > 0 && !this.reconnectId) {
             this.chatToken = optionalParams.liveChatContext.chatToken || {};
             this.requestId = optionalParams.liveChatContext.requestId || uuidv4();
 
@@ -366,6 +366,37 @@ class OmnichannelChatSDK {
 
                 console.error(`Unable to join conversation that's in '${conversationDetails.state}' state`);
                 throw Error(exceptionDetails.response);
+            }
+        }
+
+        if (this.authSettings) {
+            if (!this.authenticatedUserToken) {
+                await this.setAuthTokenProvider(this.chatSDKConfig.getAuthToken);
+            }
+
+            if (optionalParams.liveChatContext && Object.keys(optionalParams.liveChatContext).length > 0) {
+                this.chatToken = optionalParams.liveChatContext.chatToken || {};
+                this.requestId = optionalParams.liveChatContext.requestId || uuidv4();
+
+                try {
+                    await this.OCClient.validateAuthChatRecord(this.requestId, {
+                        authenticatedUserToken: this.authenticatedUserToken,
+                        chatId: this.chatToken.chatId
+                    });
+                } catch {
+                    const exceptionDetails = {
+                        response: "OCClientValidateAuthChatRecordFailed",
+                        message: "InvalidAuthChatRecord"
+                    };
+
+                    this.scenarioMarker.failScenario(TelemetryEvent.StartChat, {
+                        RequestId: this.requestId,
+                        ChatId: this.chatToken.chatId as string,
+                        ExceptionDetails: JSON.stringify(exceptionDetails)
+                    });
+
+                    throw Error(exceptionDetails.response);
+                }
             }
         }
 
@@ -445,6 +476,11 @@ class OmnichannelChatSDK {
 
         if (this.authenticatedUserToken) {
             sessionInitOptionalParams.authenticatedUserToken = this.authenticatedUserToken;
+        }
+
+        if (this.chatToken.chatId) {
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            (sessionInitOptionalParams as any).initContext.chatId = this.chatToken.chatId;
         }
 
         // Skip session init when there's a valid live chat context